Remove DataRangeWidget.py add widgets to set the range of DataGrid in DataGridWidget.py
