+++ src/UI/CustomWidgets/DataPreparation/__init__.py
@@ -1,4 +1,3 @@
 from .DataGridWidget import DataGridWidget
 from .DataPlotWidget import DataPlotWidget
 from .SelectModelColumnsWidget import SelectModelColumnsWidget
-from .DataRangeWidget import DataRangeWidget
